% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_icv_images.R
\name{read_icv_images}
\alias{read_icv_images}
\title{Lee imágenes de un directorio con extensión común.}
\usage{
read_icv_images(img_folder)
}
\arguments{
\item{img_folder}{Ruta al directorio que contiene las imágenes.}
}
\value{
Una lista de rutas de archivos de las imágenes encontradas.
}
\description{
Esta función toma una ruta de directorio y busca todas las imágenes con extensiones comunes (JPEG, JPG, PNG, BMP, TIFF, TIF) en el directorio y sus subdirectorios. Devuelve una lista de rutas de archivos de las imágenes encontradas.
}
