% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icv_grid.R
\name{add_icvgrid}
\alias{add_icvgrid}
\title{Añade una grilla a una imagen para el cálculo del índice de contaminación visual.}
\usage{
add_icvgrid(img_path, dpi = 150, save = TRUE)
}
\arguments{
\item{img_path}{Ruta al archivo de imagen.}

\item{dpi}{dpi resolucion empleada para guardar la imagen.}

\item{save}{argumento que define si la imagen se guarda o visualiza, por defecto TRUE.}
}
\value{
La función guarda la imagen con la grilla añadida en un archivo con el nombre modificado.
}
\description{
Esta función toma una imagen y agrega una grilla con líneas de diferentes colores para ayudar en el cálculo del índice de contaminación visual.
}
\examples{
# Ejemplo de uso:
# add_icvgrid("imagen.jpg")
}
